fix everhtying
<3 tilemap, camera z,
