+++ Assets/Scripts/Perks/HealthPerk.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class HealthPerk : IPerk 
 {
     public float healthToAdd;
-    public override void Apply(GameObject aGameObject)
+    public override void Apply(Player aPlayer)
     {
         // Logic to add player's health 
         Debug.Log("Apply Health Perk");
