+++ Assets/Scripts/Perks/IPerk.cs
@@ -5,5 +5,5 @@ public abstract class IPerk : ScriptableObject
 {
     public string perkName;
     public Image perkImage;
-    public abstract void Apply(GameObject aGameObject);
+    public abstract void Apply(Player aPlayer);
 }
