+++ Assets/Scripts/XPDrop.cs
@@ -17,33 +17,33 @@ public class XPDrop : MonoBehaviour
 }
 
 // Player Script (XP and Leveling Up)
-public class Player : MonoBehaviour
-{
-    private int currentXP = 0;
-    private int xpForNextLevel = 100;
-    private int playerLevel = 1;
+//public class Player : MonoBehaviour
+//{
+//private int currentXP = 0;
+//private int xpForNextLevel = 100;
+//private int playerLevel = 1;
 
-    public void AddXP(int xp)
-    {
-        currentXP += xp;
-        if (currentXP >= xpForNextLevel)
-        {
-            LevelUp();
-        }
-    }
+//public void AddXP(int xp)
+//{
+//    currentXP += xp;
+//    if (currentXP >= xpForNextLevel)
+//    {
+//        LevelUp();
+//    }
+//}
 
-    private void LevelUp()
-    {
-        playerLevel++;
-        currentXP = 0; // Reset XP
-        xpForNextLevel += 50; // Increase XP requirement for the next level
-        UpgradeStats(); // Increase player stats like health, speed, damage, etc.
-    }
+//private void LevelUp()
+//{
+//    playerLevel++;
+//    currentXP = 0; // Reset XP
+//    xpForNextLevel += 50; // Increase XP requirement for the next level
+//    UpgradeStats(); // Increase player stats like health, speed, damage, etc.
+//}
 
-    private void UpgradeStats()
-    {
-        // Example stat upgrades
-        moveSpeed += 1f;
-        // Upgrade health, damage, etc.
-    }
-}
+//private void UpgradeStats()
+//{
+//    // Example stat upgrades
+//    moveSpeed += 1f;
+//    // Upgrade health, damage, etc.
+//}
+//}
