--- Assets/Sprites/temp.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 77a70c2c2ecbb984f892295f57d70691
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
