+++ ProjectSettings/TagManager.asset
@@ -38,6 +38,18 @@ TagManager:
   - 
   - 
   m_SortingLayers:
+  - name: background
+    uniqueID: 2219944287
+    locked: 0
   - name: Default
     uniqueID: 0
+    locked: 0
+  - name: player
+    uniqueID: 3358364919
+    locked: 0
+  - name: foreground
+    uniqueID: 2921204965
+    locked: 0
+  - name: ui
+    uniqueID: 954578261
     locked: 0
